% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qol.R
\name{miss_qol}
\alias{miss_qol}
\title{miss_qol}
\usage{
miss_qol(y)
}
\arguments{
\item{x}{A data frame with ID, Q1, Q2,..., Q30 columns along with other columns if data is available.}
}
\value{
A data frame by replacing the columns 'Q1','Q2',...,'Q30' by the domain-based scale scores.
}
\description{
Creates a dataset containing the domain-based scale scores using
the data from Quality of Life questionnaire
}
\details{
Calculates the domain-based scale scores using the data from
Quality of Life questionnaire

miss_qol function inputs a dataset containing missing information, represented as,
9 or 99 or NA. It extracts only the columns named 'Q1','Q2',...,'Q30' and replaces
the missing data with the minimum value of the particular question.

Using each of the 30 columns, the Raw Score is computed, and one column is obtained containing
the Raw Score for each patient.

Further, using each of the Raw Scores, three domain-based Scale Scores are computed,
they are, Global Scales Score, Functional Scales Score and Symptoms Scales Score.

Thus, the columns 'Q1','Q2',...,'Q30' are replaced by the domain-based scale scores,
which is obtained as the output.

miss_qol(x)

1) Subject ID column should be named as 'ID'.

2) Each question column should be named as 'Q1' for data from question 1,
'Q2' for data from question 2, and so on until 'Q30' for data from question 30.

3) Data may contain more variables, such as, Age, Gender, etc.

x - A data frame with ID, Q1, Q2,..., Q30 columns along with other columns if data
is available.

rs - A matrix containing the Raw Score computed using all Q1 to Q30 data for each
patient. The RS(a) function is used in this case.

fs - A matrix containing the Functional Scale Scores computed using all Q1 to Q30
data for each patient. The FS(a,b) function is used in this case.

ss_gs - A matrix containing the Global Scale Scores computed using all Q1 to Q30
data for each patient. The SS_GS(a,b) function is used in this case.

final_data - A data frame formed by replacing the columns 'Q1','Q2',...,'Q30' by
the domain-based scale scores.
}
\examples{
##
data("sim_x")
miss_qol(sim_x)
##

}
\references{

}
\seealso{

}
\author{
Atanu Bhattacharjee and Ankita Pal
}
